#607 Extract the test Thrift struct builder
ColumnIndexReaderTest carried a private hand-rolled Thrift Compact Protocol
builder. The upcoming readers for SizeStatistics and the NaN count fields need
the same scaffolding to construct their inputs, so it moves to a shared
package-private test class rather than being copied per test class.
