+++ core/src/test/java/dev/hardwood/internal/thrift/ColumnIndexReaderTest.java
@@ -9,24 +9,16 @@ package dev.hardwood.internal.thrift;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
 
 import org.junit.jupiter.api.Test;
 
 import dev.hardwood.metadata.ColumnIndex;
 
+import static dev.hardwood.internal.thrift.ThriftStructBuilder.TYPE_I32;
+import static dev.hardwood.internal.thrift.ThriftStructBuilder.TYPE_LIST;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class ColumnIndexReaderTest {
-
-    private static final byte TYPE_I32 = 0x05;
-    private static final byte TYPE_LIST = 0x09;
-
-    // Thrift Compact Protocol list element type codes.
-    private static final byte ELEM_BOOL = 0x01;
-    private static final byte ELEM_I64 = 0x06;
-    private static final byte ELEM_BINARY = 0x08;
-    private static final byte ELEM_STRUCT = 0x0C;
 
     @Test
     void readsCoreFieldsAndSkipsHistogramAndNanCountFields() throws IOException {
@@ -75,8 +67,8 @@ class ColumnIndexReaderTest {
         assertThat(index.getPageCount()).isEqualTo(2);
     }
 
-    private static ThriftBuilder struct() {
-        return new ThriftBuilder();
+    private static ThriftStructBuilder struct() {
+        return new ThriftStructBuilder();
     }
 
     private static byte[] bytes(int... values) {
@@ -85,97 +77,5 @@ class ColumnIndexReaderTest {
             out[i] = (byte) values[i];
         }
         return out;
-    }
-
-    /// Hand-rolled Thrift Compact Protocol struct builder for tests.
-    private static final class ThriftBuilder {
-        private final ByteBuffer buffer = ByteBuffer.allocate(512).order(ByteOrder.LITTLE_ENDIAN);
-        private short lastFieldId;
-
-        ThriftBuilder field(int id, byte type) {
-            short delta = (short) (id - lastFieldId);
-            if (delta > 0 && delta <= 15) {
-                buffer.put((byte) ((delta << 4) | (type & 0x0F)));
-            }
-            else {
-                buffer.put(type);
-                writeZigzag(id);
-            }
-            lastFieldId = (short) id;
-            return this;
-        }
-
-        ThriftBuilder boolList(boolean... values) {
-            listHeader(values.length, ELEM_BOOL);
-            for (boolean value : values) {
-                buffer.put((byte) (value ? 0x01 : 0x02));
-            }
-            return this;
-        }
-
-        ThriftBuilder binaryList(byte[]... values) {
-            listHeader(values.length, ELEM_BINARY);
-            for (byte[] value : values) {
-                writeVarint(value.length);
-                buffer.put(value);
-            }
-            return this;
-        }
-
-        ThriftBuilder i64List(long... values) {
-            listHeader(values.length, ELEM_I64);
-            for (long value : values) {
-                writeZigzag(value);
-            }
-            return this;
-        }
-
-        ThriftBuilder emptyStructList(int count) {
-            listHeader(count, ELEM_STRUCT);
-            for (int i = 0; i < count; i++) {
-                buffer.put((byte) 0); // empty struct: immediate STOP
-            }
-            return this;
-        }
-
-        ThriftBuilder i32(int value) {
-            writeZigzag(value);
-            return this;
-        }
-
-        ThriftBuilder stop() {
-            buffer.put((byte) 0);
-            return this;
-        }
-
-        byte[] build() {
-            byte[] out = new byte[buffer.position()];
-            buffer.flip();
-            buffer.get(out);
-            return out;
-        }
-
-        private void listHeader(int size, byte elementType) {
-            if (size < 15) {
-                buffer.put((byte) ((size << 4) | (elementType & 0x0F)));
-            }
-            else {
-                buffer.put((byte) (0xF0 | (elementType & 0x0F)));
-                writeVarint(size);
-            }
-        }
-
-        private void writeVarint(long value) {
-            long v = value;
-            while ((v & ~0x7FL) != 0) {
-                buffer.put((byte) ((v & 0x7F) | 0x80));
-                v >>>= 7;
-            }
-            buffer.put((byte) (v & 0x7F));
-        }
-
-        private void writeZigzag(long value) {
-            writeVarint((value << 1) ^ (value >> 63));
-        }
     }
 }
